requirement.txt
streamlit==1.33.0
langchain-openai==0.1.13
langchain-core==0.2.10
langchain-prompts==0.2.10
langchain==0.2.6
langchain-community==0.2.6
python-dotenv==1.0.1
requests==2.31.0
